Tests of setProviderInfo

+++ contracts/CustomOracle.sol
@@ -6,11 +6,11 @@ pragma solidity ^0.8.4;
 
 contract CustomOracle{
     address public owner; 
-    mapping(uint => uint) public expiryDatePrice; 
-    mapping(uint => bool) public expiryAvailable; 
-    constructor(uint[] memory _expiryDates){
+    mapping(uint256 => uint256) public expiryDatePrice; 
+    mapping(uint256 => bool) public expiryAvailable; 
+    constructor(uint256[] memory _expiryDates){
         owner = msg.sender; 
-        for(uint i=0; i<_expiryDates.length; i++){
+        for(uint256 i=0; i<_expiryDates.length; i++){
             expiryAvailable[_expiryDates[i]] = true; 
         }
     }
@@ -20,7 +20,7 @@ contract CustomOracle{
         _;
     }
 
-    function setExpiryPrice(uint _expiry, uint _price) public onlyOwner {
+    function setExpiryPrice(uint256 _expiry, uint256 _price) public onlyOwner {
         require(expiryAvailable[_expiry] == true, "No contract with this expiry."); 
         require(block.timestamp > _expiry, "Expiry !< currentTime.");
         require(expiryDatePrice[_expiry] == 0, "Data already set.");
@@ -28,13 +28,13 @@ contract CustomOracle{
         return; 
     }
 
-    function getExpiryPriceLevel(uint _expiry) public view returns(uint){
+    function getExpiryPriceLevel(uint256 _expiry) public view returns(uint256){
         require(expiryAvailable[_expiry] == true, "No contract with this expiry."); 
         require(expiryDatePrice[_expiry] != 0, "Data is still unavailable"); 
         return expiryDatePrice[_expiry]; 
     }
 
-    function makeCorrection(uint _expiry, uint _price) public onlyOwner{
+    function makeCorrection(uint256 _expiry, uint256 _price) public onlyOwner{
         require(expiryAvailable[_expiry] == true, "No contract with this expiry."); 
         require(expiryDatePrice[_expiry] != 0, "Data is not set.");
         expiryDatePrice[_expiry] = _price;
